Use .quiet in Response::Mashify
The use of `Hashie::Mash` in `Response::Mashify` creates a lot of warnings
per response in the logs that aren't really valuable, and similar projects ignore
them as well.

`Hashie::Mash.quiet` is a built-in class method that creates an anonymous
class to silence the warnings it creates.

```ruby
irb> Hashie::Mash.new({"size" => "foo"})
You are setting a key that conflicts with a built-in method Hashie::Mash#size defined in Hash. This can cause unexpected behavior when accessing the key as a property. You can still access the key via the #[] method.
#=> {"size"=>"foo"}
irb> Hashie::Mash.quiet.new({"size" => "foo"})
#=> {"size"=>"foo"}
```